Move Translation Key To Code Storage to A Def
Create the LegacuReasonToDetailOptionsDef which is just meant to be an XML-configurable mapping between translation keys for various legacy flavors to their selection in code. Instead of manually selection translation keys A,B,C for a specific legacy reason, we can now just select from the relevant list in the one-off mapping def.