Refactor Mythic Comp to only get added to real mythic items
Removed code that adds mythic comp to all apparel and weapons on startup, and replaced it with code which saves a flag to the comp's parent thing, and uses that flag to reload the comp upon loading a game, even though the comp isn't part of the thingDef's normal list of comps.

Also added a few descriptions, which inlcude the 'M' prefix before the number to avoid name matching with HI19's contributions.